% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{class}
\name{ld_project-class}
\alias{ld_project-class}
\title{An S4 class to represent the longitudinal data after projection}
\description{
It's inherit from Longitudinal data type
}
\section{Slots}{

\describe{
\item{\code{s}}{numeric type for selected id}

\item{\code{v}}{numeric type for selected visit}

\item{\code{r}}{character type for selected room}
}}

