% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{class}
\name{ld_summary-class}
\alias{ld_summary-class}
\title{An S4 class to represent the summary of data}
\description{
An S4 class to represent the summary of data
}
\section{Slots}{

\describe{
\item{\code{summary_array}}{numeric vector for summary statistics}

\item{\code{summary_table}}{a data frame for summary data}
}}

