% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{class}
\name{LongitudinalData-class}
\alias{LongitudinalData-class}
\title{An S4 class to represent the longitudinal data}
\description{
An S4 class to represent the longitudinal data
}
\section{Slots}{

\describe{
\item{\code{id}}{numeric type}

\item{\code{visit}}{numeric type}

\item{\code{room}}{character type}

\item{\code{value}}{numeric type}

\item{\code{timepoint}}{numeric type}
}}

