% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{make_LD}
\alias{make_LD}
\title{A method to make a LongitudinalData class from given data}
\usage{
make_LD(data)
}
\arguments{
\item{data}{a data that have `id`, `visit`, `room`, `value` and `timepoint`}
}
\description{
A method to make a LongitudinalData class from given data
}
