% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{print,ld_project-method}
\alias{print,ld_project-method}
\title{A print method to show the projected `id`, `visit` and `room` of a ld_project class}
\usage{
\S4method{print}{ld_project}(x, ...)
}
\arguments{
\item{x}{the ld_project class}
}
\description{
A print method to show the projected `id`, `visit` and `room` of a ld_project class
}
