% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{print,ld_summary-method}
\alias{print,ld_summary-method}
\title{A print method to show the summary of a ld_summary class}
\usage{
\S4method{print}{ld_summary}(x, ...)
}
\arguments{
\item{x}{the ld_summary class}
}
\description{
A print method to show the summary of a ld_summary class
}
