% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{print,LongitudinalData-method}
\alias{print,LongitudinalData-method}
\title{A print method to show the number of different IDs in a LongitudinalData class}
\usage{
\S4method{print}{LongitudinalData}(x, ...)
}
\arguments{
\item{x}{the LongitudinalData class}
}
\description{
A print method to show the number of different IDs in a LongitudinalData class
}
