% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{room}
\alias{room}
\title{A method to project the data into given `room`}
\usage{
room(data, room)
}
\arguments{
\item{data}{the data to be projected}

\item{room}{the given `room`}
}
\description{
A method to project the data into given `room`
}
