% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{subject}
\alias{subject}
\title{A method to project the data into given `id`}
\usage{
subject(data, id)
}
\arguments{
\item{data}{the data to be projected}

\item{id}{the given `id`}
}
\description{
A method to project the data into given `id`
}
