% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{summary,ld_project-method}
\alias{summary,ld_project-method}
\title{A summary method to summarize the longitudinal data type}
\usage{
\S4method{summary}{ld_project}(object)
}
\arguments{
\item{object}{the data to be summarize}
}
\description{
A summary method to summarize the longitudinal data type
}
