% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{visit}
\alias{visit}
\title{A method to project the data into given `visit`}
\usage{
visit(data, visit)
}
\arguments{
\item{data}{the data to be projected}

\item{id}{the given `visit`}
}
\description{
A method to project the data into given `visit`
}
